:recycle: Refactoring validations, responses and schemas.
Movin schemas inside dto validations. Update responses.
