--- src/common/schema/index.ts
@@ -1,2 +1,0 @@
-export { default as usersSchemas } from './users.schemas'
-export { default as permissionsSchemas } from './permissions.schemas'
