+++ src/interfaces/response.d.ts
@@ -1,5 +1,4 @@
 export interface DataResponse {
-  statusCode?: number
   message: string
   data: any
 }
